% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrimeFactors.R
\name{getPrimeFactors}
\alias{getPrimeFactors}
\title{Returns the prime factors of a number}
\usage{
getPrimeFactors(x)
}
\arguments{
\item{x}{An integer to generate the prime factors for}
}
\description{
\code{getPrimeFactors} returns a vector containing all prime factors of \code{x}
}
\details{
This function is just a wrapper using \code{\link{getFactors}} and 
\code{\link{isPrime}} functions. It uses \code{\link{getFactors}} to get all factors and
\code{\link{isPrime}} to return only those that are prime.
}
\examples{
getPrimeFactors(21)
getPrimeFactors(7)
}
