% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPrime.R
\name{isPrime}
\alias{isPrime}
\title{Determines if a number is a prime}
\usage{
isPrime(x)
}
\description{
\code{isPrime} returns \code{TRUE} if \code{x} is a prime, or \code{FALSE} if not
}
\details{
This algorith uses the simple approach. Given an input number n, check whether any 
prime integer m from 2 to √n evenly divides n (the division leaves no remainder). 
If n is divisible by any m then n is composite, otherwise it is prime. See 
\href{https://en.wikipedia.org/wiki/Primality_test}{Wikipedia} for further details.

@param x An integer to test for primality
}
\examples{
isPrime(10)
isPrime(7)
}
